Top Tabs - Nit fixes
