+++ Client/Frontend/Browser/TabTrayButtonExtensions.swift
@@ -1,10 +1,6 @@
-//
-//  TabButtons.swift
-//  Client
-//
-//  Created by Tyler Lacroix on 6/27/16.
-//  Copyright © 2016 Mozilla. All rights reserved.
-//
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 import UIKit
